Move playerlist management to class
